+++ tests/VectorTest.php
@@ -233,11 +233,6 @@ class VectorTest extends TestCase
 
         $this->assertInstanceOf(ColumnVector::class, $z);
         $this->assertEquals($outcome, $z->asArray());
-
-        $z = $this->a->T;
-
-        $this->assertInstanceOf(ColumnVector::class, $z);
-        $this->assertEquals($outcome, $z->asArray());
     }
 
     public function test_map()